+++ Carbon.Permissions/Functions/Get-CPermission.ps1
@@ -1,31 +1,4 @@
 
-# Leave these here so that when Get-CPermission moves to its own module, these go with it.
-Add-CTypeData -Type IO.DirectoryInfo `
-              -MemberName 'GetAccessControl' `
-              -MemberType ScriptMethod `
-              -Value {
-                    [CmdletBinding()]
-                    param(
-                        [Security.AccessControl.AccessControlSections] $IncludeSections =
-                            [Security.AccessControl.AccessControlSections]::All
-                    )
-
-                    return [IO.FileSystemAclExtensions]::GetAccessControl($this, $IncludeSections)
-                }
-
-Add-CTypeData -Type IO.FileInfo `
-              -MemberName 'GetAccessControl' `
-              -MemberType ScriptMethod `
-              -Value {
-                    [CmdletBinding()]
-                    param(
-                        [Security.AccessControl.AccessControlSections]$IncludeSections =
-                            [Security.AccessControl.AccessControlSections]::All
-                    )
-
-                    return [IO.FileSystemAclExtensions]::GetAccessControl($this, $IncludeSections)
-                }
-
 function Get-CPermission
 {
     <#
@@ -129,7 +102,7 @@ function Get-CPermission
             {
                 if( $item.PSProvider.Name -ne 'Certificate' )
                 {
-                    $item.GetAccessControl([Security.AccessControl.AccessControlSections]::Access) | Write-Output
+                    $item | Get-CAcl -IncludeSection ([AccessControlSections]::Access) | Write-Output
                     continue
                 }