+++ Carbon.Permissions/Functions/Revoke-CPermission.ps1
@@ -99,7 +99,7 @@ function Revoke-CPermission
             # When passed to Set-Acl, this causes intermittent errors.  So, we just grab the ACL portion of the security
             # descriptor. See
             # http://www.bilalaslam.com/2010/12/14/powershell-workaround-for-the-security-identifier-is-not-allowed-to-be-the-owner-of-this-object-with-set-acl/
-            $currentAcl = $item.GetAccessControl('Access')
+            $currentAcl = $item | Get-CAcl -IncludeSection ([AccessControlSections]::Access)
 
             foreach ($ruleToRemove in $rulesToRemove)
             {